Add files via upload
These datasets are old and small by today's
standards. However, a simple model is created, and then
various machine learning explainability tools and
techniques are used. At the start, it was hypothesized, using
domain knowledge that factors such as cholesterol, blood
pressure, and age would be major factors in the model for
heart disease prediction. This dataset didn't show that.
Instead, the number of major factors and aspects of ECG
results dominated. For Chronic Kidney disease prediction,
Feature Importance algorithm is used which shows the
increase in the percentage of model error when the
information about the feature is destroyed. The results
were different from what the theoretical knowledge
suggested the results would be. This sort of approach will
become increasingly important as machine learning has a
greater and greater role in health care.